requirements.txt

streamlit
opencv-python
mediapipe
pillow
python-dotenv
openai
